Do not copy the locally built provider globally
This came from the boilerplate we inherited from the Pulumi
Makefile.  It's generally a terrible idea, because quite
a few people have their GOBIN in their PATH.

In future, this will require people to explicitly set their
PATH to point to .../pulumi-svmkit/bin, in order for Pulumi
to be able to find the locally built provider.